requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
google-generativeai==0.5.4
youtube-transcript-api==0.6.1
python-dotenv==1.0.1
